debug robot

+++ src/Cell.java
@@ -1,9 +1,9 @@
 public class Cell implements Cloneable {
 
 	public Coordonnee coordonnee;
-	public int state; // 0: base 1: safe 2: onFire 3: Burned 4: unknown
+	public int state; // 0: base 1: safe 2: onFire 3: Burned
 	public boolean asHuman;
-	public double fire; // value between 0 and 1 with 0.1 steps down
+	public int fire; // value between 0 and 1 with 0.1 steps down
 	public int duration;
 	
 	public Cell(Coordonnee coordonnee, int state) {
@@ -14,7 +14,7 @@ public class Cell implements Cloneable {
 		this.duration=0;
 	}
 
-	public void set(int state, boolean asHuman, double fire, int duration) {
+	public void set(int state, boolean asHuman, int fire, int duration) {
 		this.state=state;
 		this.asHuman=asHuman;
 		this.fire=fire;
